# api/openapi-message-queue.yaml
openapi: "3.0.3"
info:
  title: GPU Telemetry Pipeline — Message Queue API
  description: |
    Custom in-memory ring-buffer message queue. Streamers publish telemetry
    data points; Collectors consume and acknowledge them.

    Run `make generate-openapi-message-queue` to regenerate this file from
    source annotations.
  version: "1.0"

servers:
  - url: http://localhost:8080
    description: Local docker-compose / kind deployment

tags:
  - name: messages
    description: Message lifecycle — publish, consume, acknowledge
  - name: operations
    description: Liveness, readiness, and metrics probes

paths:
  /messages:
    post:
      summary: Publish a telemetry message
      description: |
        Enqueues a single GPU telemetry data point. The call blocks until a
        free ring-buffer slot is available or the server-configured publish
        timeout elapses (default 10 s).
      operationId: publishMessage
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
            example:
              timestamp: "2025-07-18T20:42:34Z"
              metric_name: "DCGM_FI_DEV_GPU_UTIL"
              gpu_id: "0"
              device: "nvidia0"
              uuid: "GPU-5fd4f087-86f3-7a43-b711-4771313afc50"
              model_name: "NVIDIA H100 80GB HBM3"
              hostname: "mtv5-dgx1-hgpu-031"
              value: "72.5"
      responses:
        "201":
          description: Message accepted and enqueued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
              example:
                message_id: "550e8400-e29b-41d4-a716-446655440000"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: Queue is full — caller should retry with exponential backoff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /messages/consume:
    get:
      summary: Consume messages (long poll)
      description: |
        Long-polls for up to `long_poll_timeout` seconds. Returns up to
        `batch_size` messages leased to `consumer_id`. Returns **204 No Content**
        when the timeout elapses with no messages — the caller should re-poll
        immediately.

        Each returned message carries a `delivery_id` and a `lease_expires`
        timestamp. The consumer must acknowledge the delivery before the lease
        expires, or the message will be re-queued and re-delivered.
      operationId: consumeMessages
      tags: [messages]
      parameters:
        - name: consumer_id
          in: query
          required: true
          description: Unique identifier for this consumer instance (typically the pod hostname).
          schema:
            type: string
            example: "collector-pod-0"
        - name: batch_size
          in: query
          required: false
          description: Maximum number of messages to return. Defaults to the server-configured batch size.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 10
        - name: long_poll_timeout
          in: query
          required: false
          description: Maximum time to wait for messages (Go duration string, e.g. `30s`). Defaults to the server-configured value.
          schema:
            type: string
            example: "30s"
      responses:
        "200":
          description: One or more messages leased to the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsumeResponse"
        "204":
          description: Long-poll timeout elapsed with no messages — re-poll immediately
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /messages/ack:
    post:
      summary: Acknowledge messages
      description: |
        Marks delivered messages as successfully processed. The slots are
        returned to the free pool.

        If a `delivery_id` is unknown, expired, or owned by a different
        consumer, that ID is listed in `rejected_ids` and the response
        status is **207 Multi-Status**.
      operationId: ackMessages
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckRequest"
            example:
              consumer_id: "collector-pod-0"
              delivery_ids:
                - "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: All delivery IDs acknowledged successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResult"
        "207":
          description: Partial success — one or more IDs were rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResult"
        "400":
          $ref: "#/components/responses/BadRequest"

  /health:
    get:
      summary: Liveness probe
      description: Returns 200 while the message-queue process is alive.
      operationId: mqHealth
      tags: [operations]
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /ready:
    get:
      summary: Readiness probe
      description: |
        Returns 200 when ready to serve traffic. Includes live queue depth and
        capacity so operators can gauge load at a glance. Returns 503 during
        graceful shutdown.
      operationId: mqReady
      tags: [operations]
      responses:
        "200":
          description: Queue is ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
        "503":
          description: Queue is shutting down
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"

  /metrics:
    get:
      summary: Queue metrics
      description: |
        Returns a JSON snapshot of queue counters and live state:
        - `total_published` — cumulative messages enqueued
        - `total_acked` — cumulative messages acknowledged
        - `total_redelivered` — cumulative messages re-queued after lease expiry
        - `total_dropped` — cumulative messages dropped after exceeding max delivery attempts
        - `queue_depth` — current number of occupied slots
        - `in_flight` — current number of leased (unacknowledged) slots
        - `capacity` — total ring-buffer capacity
        - `uptime_seconds` — seconds since process start
      operationId: mqMetrics
      tags: [operations]
      responses:
        "200":
          description: Metrics snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricsSnapshot"

components:
  schemas:
    PublishRequest:
      type: object
      required: [timestamp, metric_name, uuid, value]
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2025-07-18T20:42:34Z"
        metric_name:
          type: string
          example: "DCGM_FI_DEV_GPU_UTIL"
        gpu_id:
          type: string
          example: "0"
        device:
          type: string
          example: "nvidia0"
        uuid:
          type: string
          example: "GPU-5fd4f087-86f3-7a43-b711-4771313afc50"
        model_name:
          type: string
          example: "NVIDIA H100 80GB HBM3"
        hostname:
          type: string
          example: "mtv5-dgx1-hgpu-031"
        value:
          type: string
          example: "72.5"
        labels_raw:
          type: string
          example: "DCGM_FI_DRIVER_VERSION=\"535.129.03\""

    PublishResponse:
      type: object
      required: [message_id]
      properties:
        message_id:
          type: string
          description: Stable ID assigned to this message by the queue (UUID v4).
          example: "550e8400-e29b-41d4-a716-446655440000"

    TelemetryMessage:
      type: object
      properties:
        message_id:
          type: string
        timestamp:
          type: string
          format: date-time
        metric_name:
          type: string
        gpu_id:
          type: string
        device:
          type: string
        uuid:
          type: string
        model_name:
          type: string
        hostname:
          type: string
        value:
          type: string
        labels_raw:
          type: string

    DeliveryItem:
      type: object
      required: [delivery_id, lease_expires, message]
      properties:
        delivery_id:
          type: string
          description: Unique ID for this delivery attempt. Pass this in the ack request.
          example: "550e8400-e29b-41d4-a716-446655440000"
        lease_expires:
          type: string
          format: date-time
          description: Time by which the consumer must ack or the message will be re-queued.
          example: "2025-07-18T20:43:04Z"
        message:
          $ref: "#/components/schemas/TelemetryMessage"

    ConsumeResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryItem"

    AckRequest:
      type: object
      required: [consumer_id, delivery_ids]
      properties:
        consumer_id:
          type: string
          example: "collector-pod-0"
        delivery_ids:
          type: array
          items:
            type: string
          example: ["550e8400-e29b-41d4-a716-446655440000"]

    AckResult:
      type: object
      required: [acked, rejected]
      properties:
        acked:
          type: integer
          example: 10
        rejected:
          type: integer
          example: 0
        rejected_ids:
          type: array
          items:
            type: string

    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "ok"

    ReadyResponse:
      type: object
      properties:
        status:
          type: string
          example: "ready"
        reason:
          type: string
          example: "shutting down"
        queue_depth:
          type: integer
          example: 42
        capacity:
          type: integer
          example: 10000

    MetricsSnapshot:
      type: object
      properties:
        queue_depth:
          type: integer
        capacity:
          type: integer
        in_flight:
          type: integer
        total_published:
          type: integer
          format: int64
        total_acked:
          type: integer
          format: int64
        total_redelivered:
          type: integer
          format: int64
        total_dropped:
          type: integer
          format: int64
        uptime_seconds:
          type: integer
          format: int64

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Missing or invalid request body / query parameter
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ServiceUnavailable:
      description: Queue is shutting down or unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"